requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
importlib-metadata==6.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dotenv==1.0.0
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
